+++ src/cookbook/ch_3_binding_vars.rs
@@ -15,15 +15,15 @@ impl QuotedText {
 /// eg "'Hello World!', said Ferris"
 /// lexing and parsing together
 ///
-fn parse_quoted_text(inp: &str) -> Result<(&str, QuotedText), dc::ParseError> {
+fn parse_quoted_text(inp: &str) -> Result<(&str, QuotedText), ParsingError> {
     // step 1: find out which quote char is used
-    let (c, quote) = dc::Cursor::from(inp)
+    let (c, quote) = Cursor::from(inp)
         .chars_in(1..=1, &['"', '\''])
         .parse_selection()
         .validate()?;
 
     // step 2: use the quote character to extract the text between quotes
-    let (c, text) = dc::Cursor::from(c)
+    let (c, text) = Cursor::from(c)
         .chars_not_in(0.., &[quote])
         .parse_selection()
         .chars_in(1..=1, &[quote])
@@ -33,9 +33,9 @@ fn parse_quoted_text(inp: &str) -> Result<(&str, QuotedText), dc::ParseError> {
 
 /// alternative implementation using "bind"
 ///
-fn parse_quoted_text_v2(inp: &str) -> Result<(&str, QuotedText), dc::ParseError> {
+fn parse_quoted_text_v2(inp: &str) -> Result<(&str, QuotedText), ParsingError> {
     let mut quote = char::default();
-    let (c, text) = dc::Cursor::from(inp)
+    let (c, text) = Cursor::from(inp)
         .chars_in(1..=1, &['"', '\''])
         .parse_selection()
         .bind(&mut quote) // store the quote found, to use below in the matching method-chain
@@ -52,7 +52,7 @@ mod tests {
     use test_log::test;
 
     #[test]
-    fn test_parse_quoted_text() -> Result<(), dc::ParseError> {
+    fn test_parse_quoted_text() -> Result<(), ParsingError> {
         let s = "'Hello World!', said Ferris";
         let (c, qt) = parse_quoted_text(s)?;
         assert_eq!(qt, QuotedText::new('\'', "Hello World!".to_string()));