+++ src/cookbook/ch_4_enum_handcrafted.rs
@@ -9,14 +9,14 @@ enum Color {
 }
 
 impl FromStr for Color {
-    type Err = dc::ParseError;
+    type Err = ParsingError;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         match s {
             "Red" => Ok(Self::Red),
             "Blue" => Ok(Self::Blue),
             "Green" => Ok(Self::Green),
-            _ => Err(dc::ParseError::NoMatch {
+            _ => Err(ParsingError::NoMatch {
                 action: "matching color",
                 args: "",
             }),
@@ -25,8 +25,8 @@ impl FromStr for Color {
 }
 
 /// uses the FromStr trait impl above
-fn parse_enum(s: &str) -> Result<(&str, Color), dc::ParseError> {
-    dc::Cursor::from(s)
+fn parse_enum(s: &str) -> Result<(&str, Color), ParsingError> {
+    Cursor::from(s)
         .text_alt(&["Red", "Blue", "Green"])
         .parse_selection()
         .validate()
@@ -38,7 +38,7 @@ mod tests {
     use test_log::test;
 
     #[test]
-    fn test_parse_enum() -> Result<(), dc::ParseError> {
+    fn test_parse_enum() -> Result<(), ParsingError> {
         // from_str expects the whole string to match
         assert_eq!(Color::from_str("Red")?, Color::Red);
         assert_eq!(Color::from_str("Red Arrow").is_err(), true);