+++ src/cookbook/ch_8_composition.rs
@@ -1,7 +1,7 @@
 use std::str::FromStr;
 
 use crate::cookbook::{ch_2_simple_example::Time, ch_7_alternate::parse_clock};
-use crate::prelude::dc::Cursor;
+use crate::prelude::Cursor;
 use crate::prelude::*;
 
 ///
@@ -15,10 +15,10 @@ struct TimePeriod(Time, Time);
 /// often using FromStr to composite parsers works nicely
 ///
 impl FromStr for TimePeriod {
-    type Err = dc::ParseError;
+    type Err = ParsingError;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let (_c, time1, time2) = dc::Cursor::from(s)
+        let (_c, time1, time2) = Cursor::from(s)
             .debug_context("time period")
             .chars_any(5..=5)
             .parse_selection() // uses .parse_selection::<Time>()
@@ -32,7 +32,7 @@ impl FromStr for TimePeriod {
 }
 
 /// where the data is not easily lexed (tokenized appropriately) a
-/// dc::cursor/stream approach can be used.
+/// Cursor/stream approach can be used.
 ///
 /// eg a train timetable where UK uses AM/PM and continental trains use 24hour clock
 ///
@@ -53,9 +53,9 @@ struct TrainTime {
 
 ///
 /// parse_with expects a closure/function that matches
-///  fn(dc::Cursor) -> Result<(dc::Cursor,T), dc::ParseError>
+///  fn(Cursor) -> Result<(Cursor,T), dc::ParseError>
 ///
-fn parse_traintime(c: &str) -> Result<(&str, TrainTime), dc::ParseError> {
+fn parse_traintime(c: &str) -> Result<(&str, TrainTime), ParsingError> {
     let (c, city, arr, dep) = Cursor::from(c)
         .debug_context("train_time")
         .word()
@@ -72,10 +72,10 @@ fn parse_traintime(c: &str) -> Result<(&str, TrainTime), dc::ParseError> {
     Ok((c, TrainTime { city, arr, dep }))
 }
 
-fn parse_timetable(s: &str) -> Result<Vec<TrainTime>, dc::ParseError> {
+fn parse_timetable(s: &str) -> Result<Vec<TrainTime>, ParsingError> {
     let mut vec = vec![];
     for line in s.lines() {
-        let c = dc::Cursor::from(line);
+        let c = Cursor::from(line);
         let (_c, tt) = c
             .parse_with(parse_traintime)
             .ws()
@@ -86,12 +86,12 @@ fn parse_timetable(s: &str) -> Result<Vec<TrainTime>, dc::ParseError> {
     Ok(vec)
 }
 
-fn parse_str_clock(s: &str) -> Result<(&str, Time), dc::ParseError> {
-    dc::Cursor::from(s).parse_with(parse_clock).validate()
+fn parse_str_clock(s: &str) -> Result<(&str, Time), ParsingError> {
+    Cursor::from(s).parse_with(parse_clock).validate()
 }
 
-fn parse_str_traintime(c: &str) -> Result<(&str, TrainTime), dc::ParseError> {
-    let (c, city, arr, dep) = dc::Cursor::from(c)
+fn parse_str_traintime(c: &str) -> Result<(&str, TrainTime), ParsingError> {
+    let (c, city, arr, dep) = Cursor::from(c)
         .word()
         .parse_selection()
         .ws()