+++ src/prelude.rs
@@ -1,13 +1,9 @@
 pub use crate::text_parser::{Bind, Matchable, Selectable};
-
-pub mod dc {
-    pub use crate::cursor::Cursor;
-    pub use crate::error::ParseError;
-}
-
+pub use crate::cursor::Cursor;
+pub use crate::error::ParsingError;
 
 pub mod lazy {
-    pub use crate::combo::SP;
     pub use crate::combo::Parser;
     pub use crate::combo::StrParser;
+    pub use crate::combo::SP;
 }