+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.5] 
+### Added
+- benchmarks (vs nom)
+
+### Changed
+- ParseError to ParsingError and now in prelude
+- Cursor now in prelude
+
 ## [0.0.4] 
 ### Added
 - parse_opt_with and parse_opt_selected